Prototype Scene Developed
-Combined Zach's UI w/ Tina's Enemies & Kaylnn's models to make prototype train car scene

+++ Assets/Scripts/EnemyScripts/EnemyParent.cs
@@ -44,7 +44,7 @@ public class EnemyParent : MonoBehaviour
         if (collision.gameObject.CompareTag("Player"))
         {
             PlayerController playerController = collision.gameObject.GetComponent<PlayerController>();
-            playerController.TakeDamage(dmg);
+            playerController.TakeDamage((int)dmg);
         }
     }
